requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
opencv-python==4.8.0
ultralytics==8.0.196
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0
torchvision==0.16.0
pillow>=10.0.0
tqdm==4.65.0
matplotlib>=3.7.1
eventlet==0.33.3 